Add authentication via TACACS+
Modules for nss and PAM are added to allow users to login through ssh
with their TACACS+ credentials and without the need to create local
per-user accounts.  If enabled, two new accounts tacacs0 and tacacs15
are added and a user is mapped to either one based on his pri-lvl
attribute (priv-lvl 15 is mapped to the latter and all others to the
former).

A modified sudo command allows users mapped to tacacs15 to become root
with their TACACS+ password.